lib: move multibyte.[ch] to curlx/
This file provides functions in the curlx set.
