+++ lib/socks_sspi.c
@@ -35,7 +35,7 @@
 #include "curlx/timeval.h"
 #include "socks.h"
 #include "curl_sspi.h"
-#include "curl_multibyte.h"
+#include "curlx/multibyte.h"
 #include "curlx/warnless.h"
 #include "strdup.h"
 /* The last 3 #include files should be in this order */
