+++ lib/vauth/ntlm.c
@@ -39,7 +39,7 @@
 #include "../sendf.h"
 #include "../curl_ntlm_core.h"
 #include "../curl_gethostname.h"
-#include "../curl_multibyte.h"
+#include "../curlx/multibyte.h"
 #include "../curl_md5.h"
 #include "../curlx/warnless.h"
 #include "../rand.h"
