+++ lib/vauth/spnego_gssapi.c
@@ -35,7 +35,7 @@
 #include "../curlx/base64.h"
 #include "../curl_gssapi.h"
 #include "../curlx/warnless.h"
-#include "../curl_multibyte.h"
+#include "../curlx/multibyte.h"
 #include "../sendf.h"
 
 /* The last #include files should be: */
