+++ lib/vtls/schannel.c
@@ -48,7 +48,7 @@
 #include "../strerror.h"
 #include "../select.h" /* for the socket readiness */
 #include "../inet_pton.h" /* for IP addr SNI check */
-#include "../curl_multibyte.h"
+#include "../curlx/multibyte.h"
 #include "../curlx/warnless.h"
 #include "x509asn1.h"
 #include "../curl_printf.h"
