+++ projects/generate.bat
@@ -155,7 +155,6 @@ rem
     ) else if "!var!" == "CURL_SRC_X_C_FILES" (
       call :element %1 lib "strcase.c" %3
       call :element %1 lib "nonblock.c" %3
-      call :element %1 lib "curl_multibyte.c" %3
       call :element %1 lib "version_win32.c" %3
     ) else if "!var!" == "CURL_SRC_X_H_FILES" (
       call :element %1 lib "config-win32.h" %3
@@ -163,7 +162,6 @@ rem
       call :element %1 lib "strcase.h" %3
       call :element %1 lib "nonblock.h" %3
       call :element %1 lib "curl_ctype.h" %3
-      call :element %1 lib "curl_multibyte.h" %3
       call :element %1 lib "version_win32.h" %3
     ) else if "!var!" == "CURL_LIB_C_FILES" (
       for /f "delims=" %%c in ('dir /b ..\lib\*.c') do call :element %1 lib "%%c" %3
