+++ winbuild/MakefileBuild.vc
@@ -698,7 +698,7 @@ CURL_FROM_LIBCURL=\
  $(CURL_DIROBJ)\timeval.obj \
  $(CURL_DIROBJ)\warnless.obj \
  $(CURL_DIROBJ)\curl_get_line.obj \
- $(CURL_DIROBJ)\curl_multibyte.obj \
+ $(CURL_DIROBJ)\multibyte.obj \
  $(CURL_DIROBJ)\version_win32.obj \
  $(CURL_DIROBJ)\dynbuf.obj \
  $(CURL_DIROBJ)\base64.obj
@@ -728,8 +728,8 @@ $(CURL_DIROBJ)\timeval.obj: ../lib/curlx/timeval.c
 	$(CURL_CC) $(CURL_CFLAGS) /Fo"$@" ../lib/curlx/timeval.c
 $(CURL_DIROBJ)\curl_get_line.obj: ../lib/curl_get_line.c
 	$(CURL_CC) $(CURL_CFLAGS) /Fo"$@" ../lib/curl_get_line.c
-$(CURL_DIROBJ)\curl_multibyte.obj: ../lib/curl_multibyte.c
-	$(CURL_CC) $(CURL_CFLAGS) /Fo"$@" ../lib/curl_multibyte.c
+$(CURL_DIROBJ)\multibyte.obj: ../lib/curlx/multibyte.c
+	$(CURL_CC) $(CURL_CFLAGS) /Fo"$@" ../lib/curlx/multibyte.c
 $(CURL_DIROBJ)\version_win32.obj: ../lib/version_win32.c
 	$(CURL_CC) $(CURL_CFLAGS) /Fo"$@" ../lib/version_win32.c
 $(CURL_DIROBJ)\warnless.obj: ../lib/curlx/warnless.c
